Ajustes dos services
